app/api/requirements.txt
flask==2.3.3
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-exporter-prometheus==0.45b0
prometheus-client==0.20.0